+++ src/negotiation/order.spec.ts
@@ -5,6 +5,7 @@ import {
   assetOrderToSwap,
   fromNominal,
   isNative,
+  Order,
   OrderAsset,
   orderSwapAssetMatchesForMaker,
   orderSwapMatchesForMaker
@@ -265,5 +266,85 @@ describe("Payload module", () => {
     const converted = fromNominal("PAY", "100", token);
     const expected = new BigNumber("100000000000000000000");
     expect(converted).toStrictEqual(expected);
+  });
+});
+
+const defaultOrderParams = {
+  tradingPair: "ETH-BTC",
+  id: "1234",
+  validUntil: 1234567890,
+  bid: {
+    ledger: "bitcoin",
+    asset: "bitcoin",
+    nominalAmount: "1.1"
+  },
+  ask: {
+    ledger: "ethereum",
+    asset: "ether",
+    nominalAmount: "99"
+  }
+};
+
+const defaultTakerCriteria = {
+  buy: {
+    asset: "bitcoin",
+    ledger: "bitcoin"
+  },
+  sell: {
+    asset: "ether",
+    ledger: "ethereum"
+  }
+};
+
+describe("Order", () => {
+  it("matches taker criteria", () => {
+    const order = new Order(defaultOrderParams, defaultTakerCriteria, () =>
+      Promise.resolve(undefined)
+    );
+
+    expect(order.matches()).toBeTruthy();
+  });
+
+  it("doesnt match taker criteria due to incorrect asset", () => {
+    const orderParams = {
+      tradingPair: "ETH-BTC",
+      id: "1234",
+      validUntil: 1234567890,
+      bid: {
+        ledger: "bitcoin",
+        asset: "bitcoin",
+        nominalAmount: "1.1"
+      },
+      ask: {
+        ledger: "ethereum",
+        asset: "PAY",
+        nominalAmount: "99"
+      }
+    };
+
+    const order = new Order(orderParams, defaultTakerCriteria, () =>
+      Promise.resolve(undefined)
+    );
+
+    expect(order.matches()).toBeFalsy();
+  });
+
+  it("doesnt match taker criteria if buy order amount is too low", () => {
+    const takerCriteria = {
+      buy: {
+        asset: "bitcoin",
+        ledger: "bitcoin",
+        minNominalAmount: "2"
+      },
+      sell: {
+        asset: "ether",
+        ledger: "ethereum"
+      }
+    };
+    const order = new Order(defaultOrderParams, takerCriteria, () =>
+      Promise.resolve(undefined)
+    );
+
+    expect(order.matches()).toBeFalsy();
   });
 });